+++ src/data/ReadId.cpp
@@ -17,7 +17,7 @@ ReadId::ReadId(const std::string& readName) : ReadName{readName}
         Data::ReadName rName{readName};
         MovieName = rName.MovieName().ToStdString();
         HoleNumber = static_cast<size_t>(rName.Zmw());
-        ZmwInterval = rName.QueryInterval();
+        if (!rName.IsCCS()) ZmwInterval = rName.QueryInterval();
     }
 }